zDAY 1/subsciber/requirements.txt
dnspython==2.7.0
motor==3.6.0
paho-mqtt==2.1.0
pymongo==4.9.2
python-dateutil==2.9.0.post0
six==1.16.0
